requirements.txt
asgiref==3.5.2
certifi @ file:///opt/conda/conda-bld/certifi_1655968806487/work/certifi
cffi==1.15.1
charset-normalizer==2.1.1
coreapi==2.3.3
coreschema==0.0.4
cryptography==38.0.1
Django==4.1.1
django-filter==22.1
django-multiselectfield==0.1.12
django-phonenumber-field==7.0.0
django-phonenumbers==1.0.1
django-rest-knox==4.2.0
djangorestframework==3.13.1
drf-yasg==1.21.3
idna==3.4
importlib-metadata==4.12.0
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
Markdown==3.4.1
MarkupSafe==2.1.1
packaging==21.3
phonenumbers==8.12.55
pycparser==2.21
pyparsing==3.0.9
pytz==2022.2.1
requests==2.28.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
sqlparse==0.4.2
uritemplate==4.1.1
urllib3==1.26.12
zipp==3.8.1